+++ src/OperationalInsights/OperationalInsights/Tables/CreateAzureOperationalInsightsSearchTableCommand.cs
@@ -20,5 +20,62 @@ namespace Microsoft.Azure.Commands.OperationalInsights.Tables
     [Cmdlet("Create", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "OperationalInsightsSearchTable", SupportsShouldProcess = true), OutputType(typeof(PSTable))]
     public class CreateAzureOperationalInsightsSearchTableCommand : OperationalInsightsBaseCmdlet
     {
+        [Parameter(Position = 0, ParameterSetName = ByWorkspaceName, Mandatory = true, ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The resource group name.")]
+        [ResourceGroupCompleter]
+        [ValidateNotNullOrEmpty]
+        public string ResourceGroupName { get; set; }
+
+        [Parameter(Position = 1, ParameterSetName = ByWorkspaceName, Mandatory = true, ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The name of the workspace that will contain the storage insight.")]
+        [ValidateNotNullOrEmpty]
+        public string WorkspaceName { get; set; }
+
+        [Parameter(Position = 2, Mandatory = true, ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The table name.")]
+        [ValidateNotNullOrEmpty]
+        public string TableName { get; set; }
+
+        [Parameter(Position = 3, Mandatory = false, ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The table retention in days, between 4 and 730. Setting this property to -1 will default to the workspace retention")]
+        [ValidateNotNullOrEmpty]
+        public int? RetentionInDays { get; set; }
+
+        [Parameter(Position = 4, Mandatory = false, ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The table total retention in days, between 4 and 2555. Setting this property to -1 will default to table retention.")]
+        [ValidateNotNullOrEmpty]
+        public int? TotalRetentionInDays { get; set; }
+
+        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "Search job query.")]
+        [ValidateNotNullOrEmpty]
+        public string SearchQuery { get; set; }
+
+        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "The timestamp to start the search from (UTC).")]
+        [ValidateNotNullOrEmpty]
+        public string StartSearchTime { get; set; }
+
+        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "The timestamp to end the search by (UTC).")]
+        [ValidateNotNullOrEmpty]
+        public string EndSearchTime { get; set; }
+
+        [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = "Limit the search job to return up to specified number of rows.")]
+        [ValidateNotNullOrEmpty]
+        public int? Limit { get; set; }
+
+        public override void ExecuteCmdlet()
+        {
+            var tableSetProperties = new PSSearchTable(
+                resourceGroupName: ResourceGroupName,
+                workspaceName: WorkspaceName,
+                tableName: TableName,
+                startSearchTime: StartSearchTime,
+                endSearchTime: EndSearchTime,
+                limit: Limit);
+
+            if (ShouldProcess(TableName, $"Update Table: {TableName}, in workspace: {WorkspaceName}, resource group: {ResourceGroupName}"))
+            {
+                WriteObject(OperationalInsightsClient.CreateSearchTable(tableSetProperties), true);
+            }
+        }
     }
 }