+++ src/OperationalInsights/OperationalInsights/Tables/MigrateAzureOperationalInsightsTableCommand.cs
@@ -17,7 +17,7 @@ using System.Net;
 
 namespace Microsoft.Azure.Commands.OperationalInsights.Tables
 {
-    [Cmdlet("Create", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "OperationalInsightsTable", SupportsShouldProcess = true), OutputType(typeof(bool))]
+    [Cmdlet("Migrate", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "OperationalInsightsTable", SupportsShouldProcess = true), OutputType(typeof(bool))]
     public class MigrateAzureOperationalInsightsTableCommand : OperationalInsightsBaseCmdlet
     {
         [Parameter(Position = 0, ParameterSetName = ByWorkspaceName, Mandatory = true, ValueFromPipelineByPropertyName = true,